+++ frontend/src/App.tsx
@@ -1,6 +1,5 @@
-import { BrowserRouter, Routes, Route } from 'react-router'
+import { BrowserRouter, Routes, Route, Navigate } from 'react-router'
 import { Toaster } from 'sonner'
-import { HomePage } from './pages/HomePage'
 import { JourneysPage } from './pages/JourneysPage'
 import { RoutesPage } from './pages/RoutesPage'
 
@@ -8,7 +7,7 @@ function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<HomePage />} />
+        <Route path="/" element={<Navigate to="/routes" replace />} />
         <Route path="/routes" element={<RoutesPage />} />
         <Route path="/journeys" element={<JourneysPage />} />
       </Routes>